Move methodLink out of components/common
