+++ src/pages/workspaces/workspace/Workflows.js
@@ -3,7 +3,7 @@ import { Fragment, useEffect, useImperativeHandle, useState } from 'react'
 import { a, div, h, label, span } from 'react-hyperscript-helpers'
 import * as breadcrumbs from 'src/components/breadcrumbs'
 import { useViewToggle, ViewToggleButtons } from 'src/components/CardsListToggle'
-import { ButtonOutline, ButtonPrimary, Clickable, IdContainer, Link, methodLink, PageBox, Select, spinnerOverlay } from 'src/components/common'
+import { ButtonOutline, ButtonPrimary, Clickable, IdContainer, Link, PageBox, Select, spinnerOverlay } from 'src/components/common'
 import { centeredSpinner, icon } from 'src/components/icons'
 import { DelayedSearchInput } from 'src/components/input'
 import { MarkdownViewer } from 'src/components/markdown'
@@ -21,6 +21,7 @@ import * as Utils from 'src/libs/utils'
 import { DockstoreTile, MethodCard, MethodRepoTile } from 'src/pages/library/Code'
 import DeleteWorkflowConfirmationModal from 'src/pages/workspaces/workspace/workflows/DeleteWorkflowConfirmationModal'
 import ExportWorkflowModal from 'src/pages/workspaces/workspace/workflows/ExportWorkflowModal'
+import { methodLink } from 'src/pages/workspaces/workspace/workflows/methodLink'
 import { wrapWorkspace } from 'src/pages/workspaces/workspace/WorkspaceContainer'
